Move AuthenticatorService to common modulMove AuthenticatorService to common modulee
