+++ common/src/main/java/org/jboss/da/common/auth/AuthenticatorService.java
@@ -1,4 +1,4 @@
-package org.jboss.da.communication.auth;
+package org.jboss.da.common.auth;
 
 import java.util.Optional;
